pyproject.toml
[tool.poetry]
name = "datasim"
version = "0.1.0"
description = ""
authors = ["Nagacharan Teja Tangirala <nagacharan.tangirala@tum.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
geopandas = "^0.12.2"
h5py = "^3.8.0"
mesa = "^1.2.1"
pyarrow = "^12.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
